added button to run shell in prefix
